+++ src/js/modules/FociDetectionModel/FociDetectionModel.tsx
@@ -9,6 +9,9 @@ import ErrorHint from "../../../sammie/js/ui/elements/ErrorHint";
 import CellResult from "./CellResult";
 import {cl, copyChange, copyRemove} from "../../../sammie/js/util";
 import ButtonIcon from "../../../sammie/js/ui/elements/ButtonIcon";
+import {printf} from "fast-printf";
+import {mean} from "../../util/math";
+import {Tooltip} from "@mui/material";
 
 /**PERSISTENT UI STATE DEFINITIONS*/
 const asResult = atomFamily<FociDetectionModelResult,string>({key:'foci-detection-model_result',default:null});
@@ -66,11 +69,34 @@ const FociDetectionModel:React.FC<IFociDetectionModelProps> = () => {
         setSelectedFoci(nv)
         changeFociSelection(curParams,curStep,nv)
     }
-    
+    const fcc = includedCells.map((cidx)=>result.selection[cidx]).map((fociInCell)=>fociInCell.length);
+    const numCells = includedCells.length;
+    const numCellsWithFoci = fcc.filter(f=>f>0).length
+    const meanFociPerCell = mean(fcc)
 	return (<div className={'foci-detection-model margin-100-neg pad-100 ' + cl(modKeys['1'],'mod-1') + cl(modKeys['2'],'mod-2') + cl(curParams.showoutlines, 'show-outlines')}>
 	    {error && <ErrorHint error={error}/> }
         {!error && result &&
             <>
+                <div className="foci-detection-model__box pad-100-excepttop pad-50-top margin-100-bottom">
+                    <Tooltip title={`Total number of included cells`} placement={'bottom'}>
+                        <div className="box-row">
+                            <span>Number of Cells:</span>
+                            <span>{numCells}</span>
+                        </div>
+                    </Tooltip>
+                    <Tooltip title={`Number of included cells that have foci absolute and in %`} placement={'bottom'}>
+                        <div className="box-row">
+                            <span>Cells with Foci:</span>
+                            <span>{numCellsWithFoci} ({printf('%.2f %%',100*numCellsWithFoci/numCells)})</span>
+                        </div>
+                    </Tooltip>
+                    <Tooltip title={`Mean number of foci over all included cells.`} placement={'bottom'}>
+                        <div className="box-row">
+                            <span>Mean Foci per Cell:</span>
+                            <span>{printf('%.2f',meanFociPerCell)}</span>
+                        </div>
+                    </Tooltip>
+                </div>
                 {curBatch.batchParameters['cellstoprocess'] < 100 &&
                 <div className="incomplete-hint col-error pad-100-bottom">Showing only first {curBatch.batchParameters['cellstoprocess']}% of dataset as preview. Change parameter in Data Input if not desired.</div>
                 }